Move subroutines from helper to helper_show
